pyproject.toml
[project]
name = "fairytale-ai-server"
version = "0.1.0"
description = ""
authors = [
    {name = "Dawid Mastalerz",email = "dawid.mastalerz27@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.1.4,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "djangorestframework-simplejwt (>=5.4.0,<6.0.0)",
    "django-extensions (>=3.2.3,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"

